requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
python-dotenv>=1.0.0
httpx>=0.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
google-generativeai>=0.5.2